Split models into multiple files.
Fixes #39

   - Rename "base" to "core".
   - Split namespace and permissions into its own file.
   - Move standard object models into "resources".
   - Clean up __init__.py.
