+++ docs/model.md
@@ -9,9 +9,33 @@
     :show-inheritance:
     :exclude-members: __str__
 
-## Base models
+## Core models
 
-::: hubuum.models.base
+Core models for hubuum. These are the foundation that permissions and resources are built on.
+
+::: hubuum.models.core
+    :members:
+    :undoc-members:
+    :fields:
+    :show-inheritance:
+    :exclude-members: __str__
+
+## Permissions
+
+Models related to the permissions system.
+
+::: hubuum.models.permissions
+    :members:
+    :undoc-members:
+    :fields:
+    :show-inheritance:
+    :exclude-members: __str__
+
+## Resources models
+
+These are the main user-facing models. They are used to create the object one will work on day to day.
+
+::: hubuum.models.resources
     :members:
     :undoc-members:
     :fields:
