+++ hubuum/tests/test_10_direct_host_model.py
@@ -1,5 +1,5 @@
 """Test module for the Host model."""
-from hubuum.models.base import Host
+from hubuum.models.resources import Host
 
 from .base import HubuumModelTestCase
