+++ hubuum/tests/test_14_direct_vendor_model.py
@@ -1,5 +1,5 @@
 """Test module for the Vendor model."""
-from hubuum.models.base import Vendor
+from hubuum.models.resources import Vendor
 
 from .base import HubuumModelTestCase
